Update calculations to use high precision floating point
- Change Segment target_speed_counts_per_hour and distance_counts from long to double
- Update kphToCountsPerHour() to return double instead of long
- Add countsToMeters() for high precision distance calculations
- Update calculateAheadBehind() to use double for target counts
- Config file now saves segment values with 6 decimal places
- Update ETA calculation in driver display to use high precision
- Update all tests to use ASSERT_NEAR for floating point comparisons
- Update Design.md to document high precision calculations

+++ rally_types.h
@@ -13,10 +13,10 @@ class I2CCounter;
 class RallyState;
 class CounterPoller;
 
-// Segment structure
+// Segment structure (high precision)
 struct Segment {
-    long target_speed_counts_per_hour = 0;  // counts per hour
-    long distance_counts = 0;               // distance in counts
+    double target_speed_counts_per_hour = 0.0;  // counts per hour (high precision)
+    double distance_counts = 0.0;                // distance in counts (high precision)
     bool autoNext = false;
 };
